requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pymysql==1.1.1
cryptography==45.0.7
python-dotenv==1.0.1
